+++ src/lib/components/ProjectCard.svelte
@@ -8,7 +8,6 @@
   import ShareProjectModal from './ShareProjectModal.svelte'
   import EditProjectForm from './EditProjectForm.svelte'
   import ConfirmationModal from './layout/ConfirmationModal.svelte'
-  import { graphStackState } from '../stores/graphStack.svelte'
 
   interface Project {
     id: number
@@ -78,7 +77,6 @@
    */
   const handleLoad = async () => {
     try {
-      graphStackState.reset()
       const projectData = await getProject(project.id)
       const { invalidEdges, registeredNetworkNodes } =
         await importGraphFromProtocol(projectData.graph)